Refactor logging API
