+++ Source/AwsCommonRuntimeKit/crt/Logger.swift
@@ -2,6 +2,28 @@
 //  SPDX-License-Identifier: Apache-2.0.
 
 import AwsCCommon
+var logger = aws_logger()
+
+public struct Logger {
+    public static func initilize(pipe: UnsafeMutablePointer<FILE>?, level: LogLevel) {
+        var options = aws_logger_standard_options()
+        options.level = level.rawValue
+        options.file = pipe
+        aws_logger_init_standard(&logger, allocator.rawValue, &options)
+        aws_logger_set(&logger)
+    }
+
+    public static func initilize(filePath: String, level: LogLevel) {
+        var logger = aws_logger()
+        filePath.withCString { cFilePath in
+            var options = aws_logger_standard_options()
+            options.level = level.rawValue
+            options.filename = cFilePath
+            aws_logger_init_standard(&logger, allocator.rawValue, &options)
+            aws_logger_set(&logger)
+        }
+    }
+}
 
 public enum LogLevel {
     case none
